+++ src/features/form/layout/utils/makeIndexedId.ts
@@ -1,4 +1,4 @@
-import { getRepeatingBinding, isRepeatingComponent } from 'src/features/form/layout/repeating';
+import { getRepeatingBinding, isRepeatingComponent } from 'src/features/form/layout/utils/repeating';
 import type { LayoutLookups } from 'src/features/form/layout/makeLayoutLookups';
 import type { IDataModelReference } from 'src/layout/common.generated';
